blackend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-reversion==5.0.1
django-mptt==0.16.0
requests==2.31.0
django-cors-headers==4.3.1
python-dotenv==1.0.1